# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContentAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type GenerationInputFeature =
  (typeof GenerationInputFeature)[keyof typeof GenerationInputFeature];

export const GenerationInputFeature = {
  blog_post: "blog_post",
  code_explain: "code_explain",
  prompt_library: "prompt_library",
} as const;

export interface GenerationInput {
  feature: GenerationInputFeature;
  /** @minLength 1 */
  input: string;
  audience: string;
  tone: string;
  keywords?: string;
}

export interface GenerationResult {
  id: number;
  result: string;
  wordCount: number;
  seoScore: number;
  readability: string;
  toneMatch: string;
  keywordsUsed: string;
  modelUsed: string;
}

export interface Generation {
  id: number;
  feature: string;
  input: string;
  audience: string;
  tone: string;
  /** @nullable */
  keywords?: string | null;
  result: string;
  wordCount: number;
  seoScore: number;
  readability: string;
  toneMatch: string;
  keywordsUsed: string;
  modelUsed: string;
  createdAt: string;
}

export type StatsByFeature = {
  blog_post: number;
  code_explain: number;
  prompt_library: number;
};

export interface Stats {
  totalGenerations: number;
  totalWords: number;
  avgSeoScore: number;
  byFeature: StatsByFeature;
}