# lib/api-zod/src/generated/types/generation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContentAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Generation {
  id: number;
  feature: string;
  input: string;
  audience: string;
  tone: string;
  /** @nullable */
  keywords?: string | null;
  result: string;
  wordCount: number;
  seoScore: number;
  readability: string;
  toneMatch: string;
  keywordsUsed: string;
  modelUsed: string;
  createdAt: Date;
}