# lib/api-zod/src/generated/types/generationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContentAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerationInputFeature } from "./generationInputFeature";

export interface GenerationInput {
  feature: GenerationInputFeature;
  /** @minLength 1 */
  input: string;
  audience: string;
  tone: string;
  keywords?: string;
}