# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContentAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatsByFeature } from "./statsByFeature";

export interface Stats {
  totalGenerations: number;
  totalWords: number;
  avgSeoScore: number;
  byFeature: StatsByFeature;
}